clicking on a career node opens its respective accordion entry
